Added new "max party size" option to Trainer settings, under a new "General" tab. This allows users to set their trainer's maximum party size, from 1 to 6. This is to allow for the rule in the Pokemon5e PHB where early-level trainers can hold fewer than 6 Pokémon in their parties.
This maximum is respected by adding and importing Pokémon. The setting is is not yet respected for transfers from the PC, and there are a few visual artifacts like a little +(Ball) icon even when your party is "full" with the new max size.

Currently, trainers still start with a party max of 6. I am considering adding an option for this to the trainer creation flow.